splitting validator by copilot
