+++ src/main/java/edu/harvard/iq/dataverse/TermsOfUseAndAccessValidator.java
@@ -5,55 +5,30 @@
  */
 package edu.harvard.iq.dataverse;
 
-import edu.harvard.iq.dataverse.util.BundleUtil;
-import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
 /**
+ * Utility class for validating TermsOfUseAndAccess.
+ * Delegates to appropriate validator based on object type.
  *
  * @author skraffmi
  */
-public class TermsOfUseAndAccessValidator implements ConstraintValidator<ValidateTermsOfUseAndAccess, TermsOfUseAndAccess> {
+public class TermsOfUseAndAccessValidator {
 
-    @Override
-    public void initialize(ValidateTermsOfUseAndAccess constraintAnnotation) {
+    /**
+     * Validates a TermsOfAccess or TermsOfUseAndLicense object.
+     * @param value the object to validate (TermsOfAccess or TermsOfUseAndLicense)
+     * @param context the constraint validator context
+     * @return true if valid, false otherwise
+     */
+    public static boolean isTOUAValid(Object value, ConstraintValidatorContext context){
 
-    }
-
-    @Override
-    public boolean isValid(TermsOfUseAndAccess value, ConstraintValidatorContext context) {
-        
-        return isTOUAValid(value, context);
-
-    }
-    
-    public static boolean isTOUAValid(TermsOfUseAndAccess value, ConstraintValidatorContext context){
-        
-        //if part of a template it is valid
-        if (value.getTemplate() != null){
+        if (value instanceof TermsOfAccess) {
+            return TermsOfAccessValidator.isTOUAValid((TermsOfAccess) value, context);
+        } else if (value instanceof TermsOfUseAndLicense) {
+            return TermsOfUseAndLicenseValidator.isTOUAValid((TermsOfUseAndLicense) value, context);
+        } else {
             return true;
         }
-        
-         //If there are no restricted files then terms are valid 
-        if (!value.getDatasetVersion().isHasRestrictedFile()) {
-            return true;
-        }
-        /*If there are restricted files then the version
-        must allow access requests or have terms of access filled in.
-         */
-        boolean valid = value.isFileAccessRequest() == true || (value.getTermsOfAccess() != null && !value.getTermsOfAccess().isEmpty());
-        if (!valid) {
-            try {
-                if (context != null) {
-                    context.buildConstraintViolationWithTemplate(BundleUtil.getStringFromBundle("dataset.message.toua.invalid")).addConstraintViolation();
-                }
-
-               value.setValidationMessage(BundleUtil.getStringFromBundle("dataset.message.toua.invalid"));
-            } catch (NullPointerException e) {
-                return false;
-            }
-            return false;
-        }
-        return valid;
     }
 }
